Import hostname fact
This patch moves the hostname fact over from core Facter and updates the CLI
stub to call it for each defined node. Lots of things still don't work right,
but this provides a good proof-of-concept.

Concurrency in command execution has also been removed from the RFacter::Node
class for now. It is apparent that developing a fully concurrent implementation
will take some careful planning. This topic will be re-visited at a later time.