Re-enable timing of fact resolutions
This patch re-enables the measurement of fact resoltion times by passing the
`--timing` flag to rfacter or using a configuration setting. The clock used
for timing has been switched from Time.now to Process.clock_gettime which
allows a monotonic clock to be used. This isolates the timing code from events
that affect the system clock, like NTP updates