Fix spec failures related to beaker
Spec tests broke when the patch branch the Gemfile was tracking was
removed after merge. This patch switches the beaker pin to the official
puppetlabs/beaker repo and pins to the ref of the merge. Beaker releases
approximently every two weeks, so we should be able to switch back to
an official version soon.