(break) RFacter collections are now per-node
This patch updates the RFacter::Util::Collection class to require a
RFacter::Node object during initialization. Instance methods like `#value`
and `#to_hash` no longer require a node to be passed in. This means that
each collection instance is now bound to a single node and multiple
collections must be created to collect facts from multiple nodes.

Some benefits of this change:

  - The question of how to cache facts per-node has a simple answer:
    The existing mechanisms for fact caching are used.

  - The instance methods of the Collection class are now api-compatible with
    Facter 2.x as the added node argument is now set at initialization.

A potential downside:

  - Fact definitions will have to be loaded and evaluated for each node
    rather than once for a collection that serves multiple nodes.
    The loader will be instrumeted so that the overhead associated with
    this can be measured.