Update Beaker dependency in Gemfile
This patch updates the Beaker dependency to version 3.18.0 instead of pinning
to a specific SHA in the GitHub repo. This version shipped with support for
Docker.kill and disablement of reverse DNS for SSH daemons running inside of
containers. Both of these changes result in significant performance
improvements in the acceptance suite.